requirements.txt
requests
beautifulsoup4
colorama
urllib3
pyyaml
rich
tqdm
python-whois
mmh3
dnspython
